+++ packages/plugin-docs-cli/CHANGELOG.md
@@ -1,4 +1,25 @@
-# v0.0.12 (Tue May 26 2026)
+# Changelog
+
+## [0.2.0](https://github.com/grafana/plugin-tools/compare/@grafana/plugin-docs-cli@0.1.0...@grafana/plugin-docs-cli@0.2.0) (2026-07-09)
+
+
+### Features
+
+* **plugin-docs-cli:** Extract TOC headings at build time ([#2744](https://github.com/grafana/plugin-tools/issues/2744)) ([0a05680](https://github.com/grafana/plugin-tools/commit/0a05680e7961c52fcf68e567a5e58ffb1c4ae8e7))
+
+
+### Bug Fixes
+
+* **plugin-docs-cli:** limit doc page nesting to 3  ([#2745](https://github.com/grafana/plugin-tools/issues/2745)) ([95bb69a](https://github.com/grafana/plugin-tools/commit/95bb69a3011343948e2450db9afa6c9c14b79c1c))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @grafana/plugin-docs-parser bumped from ^0.1.0 to ^0.2.0
+
+## v0.0.12 (Tue May 26 2026)
 
 #### 🐛 Bug Fix
 
@@ -10,7 +31,7 @@
 
 ---
 
-# v0.0.11 (Tue May 26 2026)
+## v0.0.11 (Tue May 26 2026)
 
 #### 🐛 Bug Fix
 
@@ -22,7 +43,7 @@
 
 ---
 
-# v0.0.10 (Thu Mar 19 2026)
+## v0.0.10 (Thu Mar 19 2026)
 
 #### 🐛 Bug Fix
 
@@ -33,8 +54,6 @@
 - Erik Sundell ([@sunker](https://github.com/sunker))
 
 ---
-
-# @grafana/plugin-docs-cli
 
 ## 0.0.1 (Unreleased)